+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Added :**
 
 - sort-array-manually.rb [todo from chapter-7]
+- new-table-of-contents.rb [todo from chapter-7]
 
 ## [0.0.9] (May 21 2021)